Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
tensorflow-cpu>=2.12.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
python-multipart>=0.0.6
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.25.0
pandas>=1.5.0
